+++ src/util/getAllConfigs.js
@@ -11,9 +11,21 @@ export default function getAllConfigs(config) {
     // Add experimental configs here...
     respectDefaultRingColorOpacity: {
       theme: {
-        ringColor: {
+        ringColor: ({ theme }) => ({
           DEFAULT: '#3b82f67f',
-        },
+          ...theme('colors'),
+        }),
+      },
+    },
+
+    disableColorOpacityUtilitiesByDefault: {
+      corePlugins: {
+        backgroundOpacity: false,
+        borderOpacity: false,
+        divideOpacity: false,
+        placeholderOpacity: false,
+        ringOpacity: false,
+        textOpacity: false,
       },
     },
   }